% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\name{bec,nls-method}
\alias{bec,nls-method}
\alias{bec}
\alias{bec,mle-method}
\title{Calculate bout ending criteria from model coefficients}
\usage{
\S4method{bec}{nls}(fit)

\S4method{bec}{mle}(fit)
}
\arguments{
\item{fit}{Object of class \code{nls} or \code{mle}.}
}
\value{
\code{numeric} vector with the bout ending criterion or
    criteria derived from the model.
}
\description{
Calculate bout ending criteria from model coefficients
}
\section{Functions}{
\itemize{
\item \code{bec,nls-method}: Calculate BEC on \code{nls} object

\item \code{bec,mle-method}: Calculate BEC on \code{mle} object
}}

\author{
Sebastian P. Luque \email{spluque@gmail.com}
}
\keyword{manip}
\keyword{models}
