% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouts_helpers.R
\name{boutsCDF}
\alias{boutsCDF}
\title{Estimated cumulative frequency for two- or three-process Poisson
mixture models}
\usage{
boutsCDF(x, p, lambdas)
}
\arguments{
\item{x}{numeric vector described by model.}

\item{p}{numeric scalar or vector of proportion parameters.}

\item{lambdas}{numeric vector of rate parameters.}
}
\value{
numeric vector with cumulative frequency.
}
\description{
Estimated cumulative frequency for two- or three-process Poisson
mixture models
}
\examples{
utils::example("rmixexp", package="diveMove", ask=FALSE)
## boutsCDF(rndprocs3, p=p_true, lambdas=lda_true)
}
\author{
Sebastian P. Luque \email{spluque@gmail.com}
}
