% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\name{boutsNLSll,Bouts-method}
\alias{boutsNLSll,Bouts-method}
\alias{boutsNLSll}
\alias{boutsNLSll,numeric-method}
\title{Generalized log likelihood function taking any number of Poisson
    processes in a "broken-stick" model}
\usage{
\S4method{boutsNLSll}{Bouts}(obj, coefs)

\S4method{boutsNLSll}{numeric}(obj, coefs)
}
\arguments{
\item{obj}{Object of class \code{\link{Bouts}} or numeric vector of
independent data to be described by the function.}

\item{coefs}{matrix of coefficients (\code{a} and \code{lambda}) in
rows for each process of the model in columns.}
}
\value{
numeric vector as \code{x} with the evaluated function.
}
\description{
Generalized log likelihood function taking any number of Poisson
    processes in a "broken-stick" model
}
\section{Methods (by class)}{
\itemize{
\item \code{Bouts}: Log likelihood \code{Bouts} method

\item \code{numeric}: Log likelihood function \code{numeric} method
}}

\author{
Sebastian P. Luque \email{spluque@gmail.com}
}
