% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diveMove-defunct.R
\name{diveMove-defunct}
\alias{diveMove-defunct}
\alias{bouts2.ll}
\alias{bouts2.LL}
\alias{bouts.mle}
\title{Defunct functions in package \sQuote{diveMove}}
\usage{
bouts2.ll()

bouts2.LL()

bouts.mle()
}
\description{
These functions are defunct and no longer available.
}
\section{\code{bouts2.ll} and \code{bouts2.LL}}{

These functions have been superseded by the new function generator
\code{\link{boutsMLEll.chooser}}
}

\keyword{internal}
