% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\name{plotBoutsCDF,nls,numeric-method}
\alias{plotBoutsCDF,nls,numeric-method}
\alias{plotBoutsCDF}
\alias{plotBoutsCDF,nls,Bouts-method}
\alias{plotBoutsCDF,mle,numeric-method}
\alias{plotBoutsCDF,mle,Bouts-method}
\title{Plot empirical and deterministic cumulative frequency distribution
Poisson mixture data and model}
\usage{
\S4method{plotBoutsCDF}{nls,numeric}(fit, obj, xlim, draw.bec = FALSE, bec.lty = 2, ...)

\S4method{plotBoutsCDF}{nls,Bouts}(fit, obj, xlim, draw.bec = FALSE, bec.lty = 2, ...)

\S4method{plotBoutsCDF}{mle,numeric}(fit, obj, xlim, draw.bec = FALSE, bec.lty = 2, ...)

\S4method{plotBoutsCDF}{mle,Bouts}(fit, obj, xlim, draw.bec = FALSE, bec.lty = 2, ...)
}
\arguments{
\item{fit}{Object of class \code{nls} or \code{mle}.}

\item{obj}{Object of class \code{\link{Bouts}}.}

\item{xlim}{2-length vector with limits for the x axis.  If omitted, a
sensible default is calculated.}

\item{draw.bec}{logical; whether to draw the BECs}

\item{bec.lty}{Line type specification for drawing the BEC reference
line.}

\item{...}{Arguments passed to \code{\link{plot.default}}.}
}
\description{
Plot empirical and deterministic cumulative frequency distribution
Poisson mixture data and model
}
\section{Methods (by class)}{
\itemize{
\item \code{fit = nls,obj = numeric}: Plot (E)CDF on \code{\link{nls}} fit object
and numeric vector

\item \code{fit = nls,obj = Bouts}: Plot (E)CDF on \code{\link{nls}} fit object
and \code{\link{Bouts}} object

\item \code{fit = mle,obj = numeric}: Plot (E)CDF on numeric vector

\item \code{fit = mle,obj = Bouts}: Plot (E)CDF on \code{\link{mle}} fit object
}}

\author{
Sebastian P. Luque \email{spluque@gmail.com}
}
\keyword{methods}
\keyword{models}
\keyword{plot}
