\name{ringy}
\docType{data}

\alias{ringy}

% -------------------------------------------------------------------------

\title{Ringed Seal Ringy ARGOS Satellite Location Data}


\description{
Satellite locations of a ringed seal caught and released in New York.
}


\format{A data frame with the following information:

  \item{time}{The date and time of the location.}

  \item{class}{The ARGOS location quality classification.}

  \item{lon, lat}{x and y geographic coordinates of each location.}

}


\seealso{
  \code{\link{readLocs}}, \code{\link{distSpeed}}.
}


\source{WhaleNet Satellite Tracking Program
  \url{http://whale.wheelock.edu}.}

% -------------------------------------------------------------------------

\keyword{datasets}
